# Conflicts:
#	azure-pipelines.yml
#	project-c/Controllers/PlantsController.cs
#	project-c/DataContext.cs
#	project-c/Migrations/DataContextModelSnapshot.cs
#	project-c/Models/Plants/Plant.cs
#	project-c/Startup.cs
#	project-c/Views/Plants/Create.cshtml

+++ project-c/Controllers/PlantsController.cs
@@ -3,11 +3,17 @@ using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net.Mime;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.Data.SqlClient;
+using Microsoft.VisualStudio.Web.CodeGeneration.Contracts.Messaging;
 using project_c.Models.Plants;
+using project_c.Models.Users;
 using project_c.Services;
+
 
 namespace project_c.Controllers
 {
@@ -15,10 +21,12 @@ namespace project_c.Controllers
     {
         private readonly DataContext _context;
         private readonly UploadService _uploadService;
+        private readonly UserManager<User> _userManager;
 
-        public PlantsController(DataContext context, UploadService upload)
+        public PlantsController(DataContext context, UserManager<User> userManager, UploadService upload)
         {
             _context = context;
+            _userManager = userManager;
             _uploadService = upload;
         }
 
@@ -59,6 +67,7 @@ namespace project_c.Controllers
         [Authorize]
         public async Task<ActionResult> Create(IFormCollection form)
         {
+            
             var name = form["name"].ToString();
             var description = form["description"].ToString();
             description = char.ToUpper(description[0]) + description.Substring(1);
@@ -80,7 +89,8 @@ namespace project_c.Controllers
                     plant.Soort = Convert.ToInt32(form["filter[Soort]"]);
                     plant.Licht = Convert.ToInt32(form["filter[Licht]"]);
                     plant.Water = Convert.ToInt32(form["filter[Water]"]);
-                    
+
+                    plant.UserId = _userManager.GetUserId(User);
                     _context.Add(plant);
                     _context.SaveChanges();
                     
@@ -120,6 +130,7 @@ namespace project_c.Controllers
 
             try
             {
+
                 var plant = _context.Plants.First(p => p.PlantId == id);
 
                 plant.Name = name;
@@ -135,10 +146,25 @@ namespace project_c.Controllers
                 if (image != null)
                 {
                     plant.ImgUrl = await _uploadService.UploadImage(image);
+				}
+				
+                var plant = _context.Plants.Find(id);
+                if (_userManager.GetUserId(User) == plant.UserId || User.IsInRole("Admin"))
+                {
+                    plant.Name = name;
+                    plant.Length = Convert.ToInt32(form["length"]);
+                    plant.Description = description; 
+                    if (image != null)
+                    {
+                        plant.ImgUrl = await uploadService.UploadImage(image);
+                    }
+                    _context.Update(plant);
+                    _context.SaveChanges();
                 }
-
-                _context.Update(plant);
-                _context.SaveChanges();
+                else
+                {
+                    return Content("Your are not authorized to edit this plant");
+                }
                 return RedirectToAction(nameof(Index));
             }
             catch
@@ -155,15 +181,28 @@ namespace project_c.Controllers
         {
             try
             {
-                var plant = _context.Plants.FirstOrDefault(p => p.PlantId == id);
-                _context.Plants.Remove(plant);
-                _context.SaveChanges();
+                var plant = _context.Plants.Find(id);
+                if (_userManager.GetUserId(User) == plant.UserId || User.IsInRole("Admin"))
+                { 
+                    _context.Plants.Remove(plant);
+                    _context.SaveChanges();
+                }
+                else
+                {
+                    return Content("Your are not authorized to delete this plant");
+                }
                 return RedirectToAction(nameof(Index));
             }
             catch
             {
                 return RedirectToAction(nameof(Details));
             }
+        }
+        public ActionResult MijnPlanten()
+        {
+            var plants = from p in _context.Plants where p.UserId == _userManager.GetUserId(User) select p;
+            
+            return View(plants);
         }
     }
 }
